Feat/config database

+++ src/model/entities/Convenio.java
@@ -1,18 +1,18 @@
 package model.entities;
 
 public class Convenio extends BaseEntity {
-	int numeroCarteirinha;
+	long numeroCarteirinha;
 	String prestadora;
 	String plano;
 	
-	public Convenio(int id, int numeroCarteirinha, String prestadora, String plano) {
+	public Convenio(int id, long numeroCarteirinha, String prestadora, String plano) {
 		super(id);
 		this.numeroCarteirinha = numeroCarteirinha;
 		this.prestadora = prestadora;
 		this.plano = plano;
 	}
 	
-	public Convenio(int numeroCarteirinha, String prestadora, String plano) {
+	public Convenio(long numeroCarteirinha, String prestadora, String plano) {
 		this.numeroCarteirinha = numeroCarteirinha;
 		this.prestadora = prestadora;
 		this.plano = plano;
